add strategyQA.py for strategy retrival evalution
